Serve donate V1 only; block external /donate-v2 (#299, #292)
Pause the donate A/B test behind a single feature flag so V1 is the only
Support page that reaches production, while keeping all V2 source for a
clean re-enable.

- Add DONATE_AB_TEST_ENABLED (false) as the master switch in a shared
  donateExperiment module.
- Proxy: when disabled, serve V1 to everyone with no bucketing and no cookie;
  redirect direct/external hits to /donate-v2 back to /donate so the variant
  URL is not indexable or shareable (#292).
- Migrate the routing file from the deprecated `middleware` convention to the
  Next 16 `proxy` convention (src/middleware.ts -> src/proxy.ts), so it
  actually executes and clears the deprecation warning. The old middleware
  file was inert under Next 16, which is why the A/B test had effectively
  stopped bucketing.
- Add a getServerSideProps guard on the /donate-v2 page that redirects to
  /donate while the test is paused, as a framework-independent backstop. When
  re-enabled, the page renders normally so the internal /donate -> v2 rewrite
  keeps working.

V2 components and the V2 page are unchanged and re-enable by flipping the flag.
